UNTESTED: correct the mate and slave packets against the target
Not verified in game. Test mechanics.

The client does not treat pets and ships as one system. Mate covers pets and
mounts, Slave covers ships and land vehicles; they share unit ids, the skill
and buff layer, movement and attachment, but have separate managers, state
packets, equipment packets and lifetimes. Our own grouping assumed otherwise,
which is worth knowing before more is built on it.

Twenty-five packets already matched. Four opcodes did not:

Mount was on the wrong opcode entirely, so mounting was never received. Mount
and unmount share one wire packet - the separate unmount class collapsed into
the same form - and our unmount constant already held the right value while
being registered nowhere. Slave destruction was also wrong, and the escape and
my-slave packets were placeholders that were never registered at all.

One assignment was not taken: the mate-target request was placed on an opcode
the housing pass had already read directly from the target for its tax request.
Ours is kept; that one is call-graph evidence against a direct read.

The state packets were truncated. Their skill state carries three collections -
skills, tags and charges - and only two were being written, so the name, owner
and db id after it were read from the wrong bytes. In the slave one the owner
is a 64-bit persistent character id that the handler compares against the local
persistent id to decide whether the player controls this ship; we were sending
a 32-bit world object id, which is both short and a different number.

Ship equipment never reached the client: that packet wrote nothing at all. Its
payload is now implemented - a header with a 64-bit owner id, handle, db id and
flag, then up to three change records, each carrying the item before and after,
its source and destination, and an expiry. The count is clamped at three for a
ship and two for a mount, and the client discards the excess silently rather
than refusing, so larger changes have to be split server-side. Both incoming
requests read that owner id as 32 bits and had everything after it misaligned;
the mate reply was also missing the per-record expiry.

Attachment matched without change, including the sentinel that omits the parent
id, which confirms a passenger's position is never sent as coordinates here.

+++ AAEmu.Game/Core/Packets/C2G/CSChangeMateEquipmentPacket.cs
@@ -15,11 +15,20 @@ public class CSChangeMateEquipmentPacket : GamePacket
     {
     }
 
+    /// <summary>
+    /// Request to change a pet's or mount's equipment.
+    /// </summary>
+    /// <remarks>
+    /// Header: <c>ownerPersistentId:i64, tl:u16, mateType:u32, bts:bool, num:u8</c>. The owner
+    /// id is 64 bits, not 32 - reading it short pulled the handle and everything after it off
+    /// the wrong bytes. The client clamps the record count to two here, one fewer than for a
+    /// ship.
+    /// </remarks>
     public override void Read(PacketStream stream)
     {
-        var characterId = stream.ReadUInt32();
+        var characterId = stream.ReadInt64();
         var tl = stream.ReadUInt16(); // mate tl
-        var passengerId = stream.ReadUInt32();
+        var passengerId = stream.ReadUInt32(); // mateType, generic label in the client
         var bts = stream.ReadBoolean();
         var num = stream.ReadByte();
